+++ inc-options/minimenu.php
@@ -33,6 +33,12 @@ if ( _mw_adminimize_is_active_on_multisite() ) {
 				</tr>
 
 				<tr>
+					<td class="row-title"><a href="#self-settings"><?php esc_attr_e(
+								'Plugin Settings', 'adminimize'
+							); ?></a></td>
+				</tr>
+
+				<tr>
 					<td class="row-title"><a href="#admin_bar_options"><?php esc_attr_e(
 								'Admin Bar Back end Options', 'adminimize'
 							); ?></a></td>
@@ -133,7 +139,7 @@ if ( _mw_adminimize_is_active_on_multisite() ) {
 </div>
 
 <div id="poststuff" class="ui-sortable meta-box-sortables">
-	<div id="about" class="postbox ">
+	<div id="about" class="postbox">
 		<div class="handlediv" title="<?php esc_attr_e( 'Click to toggle' ); ?>"><br /></div>
 		<h3 class="hndle" id="about-sidebar"><?php esc_attr_e( 'About the plugin', 'adminimize' ) ?></h3>